debug: Improve debugging API
* Allow DEBUG_LOG without DEBUG_LINK
* Move debugInt() to debug.c
