+++ firmware/trezor.c
@@ -99,8 +99,11 @@ int main(void)
 
 	timer_init();
 
-#if DEBUG_LINK
+#if DEBUG_LOG
 	oledSetDebug(1);
+#endif
+
+#if DEBUG_LINK
 	storage_reset(); // wipe storage if debug link
 	storage_reset_uuid();
 	storage_commit();
